pull_request_template proposal (#1208)
* pull_request_template proposal

Here's a pull request template for us to iterate on. We may want more or less items depending on the team's preferences. We may want separate templates for front and backend changes, if they're typically separate, but it's easy enough for a dev to remove the list items that aren't relevant on PR creation.

* Adds Issue section with questions