Run the longer tests in a second stage. (#363)
* Run the longer tests in a second stage.

This way, if the first stage tests fail, the entire run will fail and
free up Travis resources for other PRs. It does add an extra ~6 minutes
of time in the golden case, but will hopefully prevent bottlenecks that
add ~30 minutes to other PRs.

* Don't cache certain directories.

We don't need to cache these anymore for Cypress and they seem to slow
down the unit tests.