Fix bug where the senior housing filter can be added twice with conflicting values.
This occurs if the user filters out senior housing properties as a
result of their answers to the eligibility questionnaire and then
subsequently uses the filter modal to include only senior housing
properties.

I fixed this by relying solely on the senior housing filter and not the
community type filter for determining the state of the filter. The
community type filter now throws an error if we try to determine the
backend filter type or value associated with it.