fix: pass waitlist + other info from backend
It seems the query builder logic doesn't return lists of the specified
entity type, so it's necessary to map the returned query result to the
entity type. Otherwise, `get` accessors don't work, since those are
only defined on the entity type, and not on the generic objects that
the query returns.