fix: create primary column on preferences
This commit adds a primary column on preferences, to get around an
issue with typeorm populating the preferences field on user entities,
even when there is no corresponding row in the user_preferences table.
This also adds more joins to the user repository query, so that all
necessary fields are present for tests to pass.